Refactor getUserId() to shared auth utility
Extracted the duplicated getUserId() implementation into a shared utility file lib/auth.ts and updated all server action files (inventory, trip, day-plan, luggage) to use it. This improves maintainability and ensures consistent user identification and Prisma user upserting logic across the application. It also fixed a syntax error in day-plan.actions.ts.
